requirements.txt
pandas==1.5.3
matplotlib==3.6.2
pillow==10.0.0
requests==2.31.0
